fix: can't search user in workspace by email in list recipients
risk: low
JIRA: F1-1049
